Implement method /a/changes.
After gerrit recives an action (like creating a review, comment, and so
on) it will send appropriate json from ssh server. Jenkins, which is
listening on the ssh channel will receive this message and do
appropriate action (for example, it will do the build/test whatever),
and sends back notification back to HTTP gerrit server to /a/changes
endpoint. There are couple of requests to that url, which will indicate
what jenkins is doing (like starting build, sending result info).